+++ module/core/former/examples/former_custom_collection.rs
